[bst-player-api] DivX Web Player integration completed

+++ trunk/bst-player-api/src/main/java/com/bramosystems/oss/player/core/client/impl/DivXPlayerImpl.java
@@ -78,7 +78,7 @@ public class DivXPlayerImpl extends JavaScriptObject {
     }
     }-*/;
 
-    public native final void setVolume(double volume) /*-{
+    public native final void setVolume(int volume) /*-{
     this.SetVolume(volume);
     }-*/;
